Move GameLocator namespace
