+++ Natsurainko.FluentCore/Launch/GameLocator/IGameLocator.cs
@@ -1,7 +1,7 @@
 ﻿using Nrk.FluentCore.Classes.Datas.Launch;
 using System.Collections.Generic;
 
-namespace Nrk.FluentCore.Launch;
+namespace Nrk.FluentCore.Launch.GameLocator;
 
 /// <summary>
 /// 游戏定位器接口
@@ -18,7 +18,7 @@ public interface IGameLocator
     /// </summary>
     /// <param name="absoluteId">游戏核心的 绝对Id</param>
     /// <returns></returns>
-    GameInfo GetGame(string absoluteId);
+    GameInfo? GetGame(string absoluteId);
 
     /// <summary>
     /// 快速遍历获取游戏核心
